requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.51.0
selenium==4.25.0
undetected-chromedriver==3.5.5
pyperclip==1.9.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
groq